fix: panic if list is empty
