+++ ui/widgets/spaces/commands.go
@@ -3,6 +3,7 @@ package spaces
 import (
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/prgrs/clickup/pkg/clickup"
+	"github.com/prgrs/clickup/ui/common"
 )
 
 type SpaceListReloadedMsg []clickup.Space
@@ -12,5 +13,16 @@ type SpaceListReadyMsg bool
 func SpaceListReadyCmd() tea.Cmd {
 	return func() tea.Msg {
 		return SpaceListReadyMsg(true)
+	}
+}
+
+func (m Model) getSpacesCmd() tea.Cmd {
+	return func() tea.Msg {
+		spaces, err := m.ctx.Api.GetSpaces(m.ctx.Config.DefaultTeam)
+		if err != nil {
+			return common.ErrMsg(err)
+		}
+
+		return SpaceListReloadedMsg(spaces)
 	}
 }
